Removed all instances where we force jax to use 64 bit, shall review properly later - updated logging messages and notebooks to finalise for main

+++ src/simulator/beam.py
@@ -22,7 +22,7 @@ class Beam:
             s0, 9 x N float: N rays with (x, y, z, vx, vy, vz) in m, m/s and amplitude, phase and polarisation (a, p, r) 
         """
 
-        self.Np = jnp.int64(Np)
+        self.Np = jnp.int32(Np)
         self.beam_size = beam_size
         self.divergence = divergence
         self.probing_direction = probing_direction
